+++ front/src/components/MediaPipeWebCam/index.style.tsx
@@ -1,0 +1,11 @@
+import styled, { css } from "styled-components";
+
+export const UserCanvas = styled.canvas`
+  position: absolute;
+  margin-left: auto;
+  margin-right: auto;
+  text-align: center;
+  z-index: 9;
+  width: 640px;
+  height: 480px;
+`;
