+++ front/src/components/page/Quiz/Game/index.style.tsx
@@ -1,16 +1,5 @@
 import styled, { css } from "styled-components";
 
-export const UserCanvas = styled.canvas`
-  position: absolute;
-  margin-left: auto;
-  margin-right: auto;
-  text-align: center;
-  z-index: 9;
-  width: 640px;
-  height: 480px;
-  /* opacity: 0.5; */
-`;
-
 export const ProblemBox = styled.div`
   width: 100%;
   height: 100vh;
